Move getParticiples helper file to helpers/internal folder
