+++ packages/yoastseo/src/languageProcessing/languages/pl/helpers/isPassiveSentencePart.js
@@ -1,4 +1,4 @@
-import getParticiples from "./getParticiples";
+import getParticiples from "./internal/getParticiples";
 import determineSentencePartIsPassive from "../../../helpers/passiveVoice/periphrastic/determineSentencePartIsPassive";
 import auxiliariesFactory from "../config/internal/auxiliaries.js";
 const auxiliaries = auxiliariesFactory();
